requirements.txt
streamlit
pandas
sqlalchemy>=2.0
psycopg2-binary
python-dotenv
openai
bcrypt
